NestSync-backend/requirements-minimal-dev.txt
# NestSync FastAPI Backend - Minimal Development Dependencies
# Focused on core functionality for Docker development environment testing

# =============================================================================
# Core Web Framework
# =============================================================================
fastapi>=0.104.1,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0

# =============================================================================
# GraphQL Support
# =============================================================================
strawberry-graphql>=0.214.0,<0.215.0
strawberry-graphql[fastapi]>=0.214.0,<0.215.0
graphql-core>=3.2.0,<4.0.0

# =============================================================================
# Database and ORM
# =============================================================================
sqlalchemy>=2.0.23,<2.1.0
alembic>=1.12.1,<1.13.0
asyncpg>=0.29.0,<0.30.0
psycopg2-binary>=2.9.9,<2.10.0

# =============================================================================
# Authentication and Security
# =============================================================================
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
pyjwt>=2.8.0,<3.0.0

# =============================================================================
# HTTP Client
# =============================================================================
httpx>=0.24.0,<0.25.0
requests>=2.31.0,<3.0.0

# =============================================================================
# Data Validation and Serialization
# =============================================================================
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# =============================================================================
# Configuration and Environment Management
# =============================================================================
python-dotenv>=1.0.0,<2.0.0

# =============================================================================
# Monitoring and Logging
# =============================================================================
loguru>=0.7.2,<1.0.0

# =============================================================================
# Canadian Compliance and Localization
# =============================================================================
pytz>=2023.3,<2024.0
python-dateutil>=2.8.2,<3.0.0

# =============================================================================
# Supabase Integration (Essential for Auth)
# =============================================================================
supabase>=2.18.0,<2.19.0
gotrue==2.5.4
postgrest>=0.16.0,<0.17.0

# =============================================================================
# Performance and Optimization
# =============================================================================
orjson>=3.9.10,<4.0.0